Create README.md
Safer Shorlines sees itself as an organization to support water sports enthusiasts around the world. In cooperation with other organisations, the best practical tips and safety rules are compiled and made available to all professional water sports operators, associations, training organisations and private users.

The foundation of “Safer Shorelines” is the result of an accident with Verena Heinz, a 20-year-old freediver. Verena was an ocean-loving, water-loving freediver who had a positive influence on countless people in her young years. Their senseless death is an opportunity to do everything possible to prevent accidents of this kind.

In addition to safety tips and the dissemination of rules of conduct in the field of water sports, “Safer Shorelines” also strives to make buoys the standard for the protection of coral reefs and for the marking of water sports routes. We are working to give legislators the urgency of water labelling and to create opportunities to implement them. With our commitment, we want to help promote safety for water sports of all kinds and to make the legal rules more transparent. Mindfulness and proper rule interpretation are part of this in order to enjoy the beauty and freedom in water sports. We hope that Verena’s legacy will lead to the prevention of similarly pointless and, above all, avoidable accidents in the future and the saving of lives.

Goal: More safety for water sports of all kinds

Location: All countries that carry out water sports activities

How: Establishment of national organisations, best practice guides and training courses